Добавлены пакеты. Добавлен файл свойств главного сервеса ewm
