--- ewm-service/src/main/java/EwmService.java
@@ -1,9 +1,0 @@
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EwmService {
-    public static void main (String[] args) {
-        SpringApplication.run(EwmService.class, args);
-    }
-}
