+++ ewm-service/src/main/java/ru/practicum/ewm/EwmService.java
@@ -1,0 +1,11 @@
+package ru.practicum.ewm;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EwmService {
+    public static void main(String[] args) {
+        SpringApplication.run(EwmService.class, args);
+    }
+}
