+++ ewm-service/src/main/resources/aplication.properties
@@ -1,5 +1,10 @@
 logging.level.org.springframework.web.client.RestTemplate=DEBUG
 
 server.port=8080
+stats-service.url=http://localhost:9090
 
-stats-service.url = http://localhost:9090
+spring.jpa.hibernate.ddl-auto=none
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL10Dialect
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.show_sql=true
+spring.sql.init.mode=always
