+++ stats/stats-client/src/main/resources/aplication.properties
@@ -1,0 +1,3 @@
+server.port=8080
+
+stats-service.url=http://localhost:9090
