--- stats/stats-dto/src/main/java/ru/practicum/ApplicationDtoForHitsInt.java
@@ -1,14 +1,0 @@
-package ru.practicum;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@RequiredArgsConstructor
-@AllArgsConstructor
-public class ApplicationDtoForHitsInt {
-    private String app;
-    private String uri;
-    private int hits;
-}
