+++ stats/stats-dto/src/main/java/ru/practicum/stats/ApplicationDtoForHitsInt.java
@@ -1,0 +1,14 @@
+package ru.practicum.stats;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@RequiredArgsConstructor
+@AllArgsConstructor
+public class ApplicationDtoForHitsInt {
+    private String app;
+    private String uri;
+    private int hits;
+}
