+++ stats/stats-dto/src/main/java/ru/practicum/stats/HitDto.java
@@ -1,0 +1,15 @@
+package ru.practicum.stats;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@RequiredArgsConstructor
+@AllArgsConstructor
+public class HitDto {
+    private String app;
+    private String uri;
+    private String ip;
+    private String timestamp;
+}
