--- stats/stats-service/src/main/java/ru/practicum/StatsService.java
@@ -1,11 +1,0 @@
-package ru.practicum;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StatsService {
-    public static void main(String[] args) {
-        SpringApplication.run(StatsService.class, args);
-    }
-}
