--- stats/stats-service/src/main/java/ru/practicum/dtoSer/ApplicationDto.java
@@ -1,12 +1,0 @@
-package ru.practicum.dtoSer;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@RequiredArgsConstructor
-@AllArgsConstructor
-public class ApplicationDto {
-    private String app;
-}
