--- stats/stats-service/src/main/java/ru/practicum/dtoSer/ApplicationDtoForHits.java
@@ -1,15 +1,0 @@
-package ru.practicum.dtoSer;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@RequiredArgsConstructor
-@AllArgsConstructor
-public class ApplicationDtoForHits {
-    private String app;
-    private String uri;
-    private Long hits;
-}
-
