--- stats/stats-service/src/main/java/ru/practicum/exceptions/DataTimeException.java
@@ -1,10 +1,0 @@
-package ru.practicum.exceptions;
-
-import lombok.Getter;
-
-@Getter
-public class DataTimeException extends RuntimeException {
-    public DataTimeException(String message) {
-        super(message);
-    }
-}
