--- stats/stats-service/src/main/java/ru/practicum/exceptions/ErrorResponse.java
@@ -1,13 +1,0 @@
-package ru.practicum.exceptions;
-
-import lombok.Getter;
-
-@Getter
-public class ErrorResponse {
-    private final String error;
-
-    public ErrorResponse(String error) {
-        this.error = error;
-    }
-}
-
