--- stats/stats-service/src/main/java/ru/practicum/mapper/HitMapper.java
@@ -1,14 +1,0 @@
-package ru.practicum.mapper;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import ru.practicum.HitDto;
-import ru.practicum.model.Hit;
-
-@Mapper(componentModel = "spring")
-public interface HitMapper {
-    @Mapping(target = "id", ignore = true)
-    @Mapping(target = "app", ignore = true)
-    @Mapping(target = "timestamp", dateFormat = "yyyy-MM-dd HH:mm:ss")
-    Hit toHit(HitDto hitDto);
-}
