--- stats/stats-service/src/main/java/ru/practicum/mapper/LongHitsToIntHits.java
@@ -1,14 +1,0 @@
-package ru.practicum.mapper;
-
-import org.mapstruct.Qualifier;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Qualifier
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.CLASS)
-public @interface LongHitsToIntHits {
-}
