--- stats/stats-service/src/main/java/ru/practicum/model/Hit.java
@@ -1,31 +1,0 @@
-package ru.practicum.model;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "hits")
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-public class Hit {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "hit_id")
-    private long id;
-    @ManyToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name = "app_id")
-    private Application app;
-    @Column(name = "uri")
-    private String uri;
-    @Column(name = "ip")
-    private String ip;
-    @Column(name = "timestamp")
-    private LocalDateTime timestamp;
-}
