--- stats/stats-service/src/main/java/ru/practicum/service/ApplicationService.java
@@ -1,11 +1,0 @@
-package ru.practicum.service;
-
-import ru.practicum.model.Application;
-
-import java.util.Optional;
-
-public interface ApplicationService {
-    Optional<Application> fineByName(String app);
-
-    Optional<Application> add(String app);
-}
