+++ stats/stats-service/src/main/java/ru/practicum/stats/StatsService.java
@@ -1,0 +1,11 @@
+package ru.practicum.stats;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StatsService {
+    public static void main(String[] args) {
+        SpringApplication.run(StatsService.class, args);
+    }
+}
