+++ stats/stats-service/src/main/java/ru/practicum/stats/dtoSer/ApplicationDto.java
@@ -1,0 +1,12 @@
+package ru.practicum.stats.dtoSer;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@RequiredArgsConstructor
+@AllArgsConstructor
+public class ApplicationDto {
+    private String app;
+}
