+++ stats/stats-service/src/main/java/ru/practicum/stats/dtoSer/ApplicationDtoForHits.java
@@ -1,0 +1,15 @@
+package ru.practicum.stats.dtoSer;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@RequiredArgsConstructor
+@AllArgsConstructor
+public class ApplicationDtoForHits {
+    private String app;
+    private String uri;
+    private Long hits;
+}
+
