+++ stats/stats-service/src/main/java/ru/practicum/stats/exceptions/DataTimeException.java
@@ -1,0 +1,10 @@
+package ru.practicum.stats.exceptions;
+
+import lombok.Getter;
+
+@Getter
+public class DataTimeException extends RuntimeException {
+    public DataTimeException(String message) {
+        super(message);
+    }
+}
