+++ stats/stats-service/src/main/java/ru/practicum/stats/exceptions/ErrorResponse.java
@@ -1,0 +1,13 @@
+package ru.practicum.stats.exceptions;
+
+import lombok.Getter;
+
+@Getter
+public class ErrorResponse {
+    private final String error;
+
+    public ErrorResponse(String error) {
+        this.error = error;
+    }
+}
+
