+++ stats/stats-service/src/main/java/ru/practicum/stats/repository/ApplicationRepository.java
@@ -1,0 +1,12 @@
+package ru.practicum.stats.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import ru.practicum.stats.model.Application;
+
+import java.util.Optional;
+
+@Repository
+public interface ApplicationRepository extends JpaRepository<Application, Long> {
+    Optional<Application> findByApp(String app);
+}
